requirements.txt
Flask
Flask-Cors
torch
torchvision
transformers
Pillow
opencv-python
numpy
matplotlib
google-generativeai
flask
python-dotenv
timm